+++ parser/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support parsing newtypes of the form `newtype IDENT = TYPE;`.
 - Add error type `RecursiveNewtype`.
+- Add warning type `NonCamelCaseNewtype`.
 
 ## [0.12.0] - 2025-01-26